requirements.txt
streamlit==1.48.1
langchain-groq==0.3.7
python-dotenv==1.1.1
langchain-core==0.3.74